feat: guess whether command was invoked non-interactively
Previously, it was made possible to control the interactivity status of the
progress bar by setting `_G.interactive`. However, this was not
automatically inferred from the context the process was started from,
and was expected to be manually set by the library invoker.

Non-interactivity status is now inferred automatically based on the
context the command was called from.