feat(lib): add support for the `.tar.xz` compression method
Some tools (currently only asphalt, as of v1.0.0-pre.16) use `.tar.xz`,
which was formerly an unimplemented (although planned) compression
method. We use the `tar` CLI, that is generally supported on macOS,
Linux and Windows as of now.

In the future, it might be a better idea to implement the compression in
Luau itself, similar to `luau-unzip` for even more portability.